django_app/requirements.txt
Django==1.11.5
djangorestframework-jsonapi==2.5.0
psycopg2
